v2.9.67 fix: physicalMemory UInt64转Int64

+++ Sources/TrollMCP2/DeviceProbe.swift
@@ -285,7 +285,7 @@ final class DeviceProbe: ObservableObject {
 
     static func memoryTotal() -> String {
         let total = ProcessInfo.processInfo.physicalMemory
-        return formatBytes(total)
+        return formatBytes(Int64(total))
     }
 
     static func workspaceSize() -> String {
